Excercise 7.15: Add view for single blog & refactor components

+++ bloglist-frontend/src/components/Blog.js
@@ -1,54 +1,42 @@
-import React, { useState } from 'react'
-import PropTypes from 'prop-types'
+import React from 'react'
+import { useDispatch, useSelector } from 'react-redux'
+import { likeBlog, removeBlog } from '../reducers/blogReducer'
+import { useHistory } from 'react-router-dom'
 
-const Blog = ({ blog, handleLike, handleRemove, user }) => {
-  const [view, setView] = useState(false)
+const Blog = ({ blog }) => {
+  const history = useHistory()
+  const dispatch = useDispatch()
+  const user = useSelector((state) => state.user)
 
-  const blogStyle = {
-    paddingTop: 10,
-    paddingLeft: 2,
-    paddingBottom: 5,
-    border: 'solid',
-    borderWidth: 1,
-    marginBottom: 5,
+  const handleLike = async (blog) => {
+    dispatch(likeBlog(blog))
   }
 
-  const fullView = () => (
+  const handleRemove = async (blog) => {
+    if (window.confirm(`Remove blog ${blog.title} by ${blog.author}?`)) {
+      dispatch(removeBlog(blog))
+      history.push("/")
+    }
+  }
+
+  if (!blog) return null
+  return (
     <div>
-      <a target="_blank" rel="noopener noreferrer" href={blog.url}>
+      <h1>{blog.title}</h1>
+      <h2>by {blog.author}</h2>
+      <a href={blog.url} target="_blank" rel="noopener noreferrer">
         {blog.url}
       </a>
-      <br />
-      likes: {blog.likes} <button id="likeBtn" onClick={() => handleLike(blog)}>like</button>
-      <br />
-      {blog.user.name}
-      <br />
+      <p>{blog.likes} likes</p>
+      <button id="likeBtn" onClick={() => handleLike(blog)}>
+        like
+      </button>
+      <p>added by {blog.user.name}</p>
       {user !== null && blog.user.name === user.name ? (
         <button onClick={() => handleRemove(blog)}>remove</button>
       ) : null}
     </div>
   )
-
-  return (
-    <div className='blog' style={blogStyle}>
-      {blog.title} {blog.author}{' '}
-      <button id="toggleView"
-        onClick={() => {
-          setView(!view)
-        }}
-      >
-        {view ? 'hide' : 'view'}
-      </button>
-      {view ? fullView() : null}
-    </div>
-  )
-}
-
-Blog.propTypes = {
-  blog: PropTypes.object.isRequired,
-  handleLike: PropTypes.func.isRequired,
-  handleRemove: PropTypes.func.isRequired,
-  user: PropTypes.object
 }
 
 export default Blog
